Change: Organize files in `compile` module
