+++ src/compile/codegen.rs
@@ -1,45 +1,44 @@
+use super::compilation_result::CompilationResult;
+use crate::expression::Expression;
+use crate::ir;
+use crate::parser;
+use crate::transform::TransformManager;
+use crate::translator::eir_translator::Builder;
+use crate::translator::{translate_ast, translate_eir};
+
 use failure::Error;
-use inkwell::targets::{CodeModel, RelocMode, Target, TargetMachine};
-use inkwell::OptimizationLevel;
 
-use super::error::LLVMError;
+use inkwell::context;
 
-pub fn create_target_machine(
-    target_triple: Option<&String>,
-    target_cpu: Option<&String>,
-    target_cpu_features: Option<&String>,
-    optimization_level: OptimizationLevel,
-    reloc_mode: RelocMode,
-) -> Result<TargetMachine, Error> {
-    let (triple, default_cpu, default_features) = if let Some(triple) = target_triple {
-        (triple.clone(), String::new(), String::new())
-    } else {
-        (
-            TargetMachine::get_default_triple().to_string(),
-            TargetMachine::get_host_cpu_name().to_string(),
-            TargetMachine::get_host_cpu_features().to_string(),
-        )
-    };
+pub fn compile_eir(eir: ir::Node, module_name: &str) -> Result<CompilationResult, Error> {
+    let context = context::Context::get_global();
+    let inst_builder = context.create_builder();
 
-    let cpu = target_cpu.unwrap_or(&default_cpu);
-    let cpu_features = target_cpu_features.unwrap_or(&default_features);
-    let target =
-        Target::from_triple(&triple).map_err(|message| LLVMError::TargetInitializationFailed {
-            message: message.to_string(),
-        })?;
-    target
-        .create_target_machine(
-            &triple,
-            &cpu,
-            &cpu_features,
-            optimization_level.into(),
-            reloc_mode.into(),
-            CodeModel::Default,
-        )
-        .ok_or_else(|| {
-            LLVMError::TargetInitializationFailed {
-                message: "Failed to create TargetMachine'".to_string(),
-            }
-            .into()
-        })
+    let module = context.create_module(module_name);
+
+    let i64_type = context.i64_type();
+    let fn_type = i64_type.fn_type(&[], false);
+
+    let function = module.add_function(module_name, fn_type, None);
+    let basic_block = context.append_basic_block(&function, "entry");
+
+    inst_builder.position_at_end(&basic_block);
+
+    let mut builder = Builder::new(inst_builder, module);
+
+    let evaluated_value = translate_eir(&mut builder, eir)?.expect_value()?;
+    builder.ret_int(evaluated_value)?;
+
+    return Ok(CompilationResult::new(builder.take_module()));
+}
+
+pub fn compile_ast(ast: Expression, module_name: &str) -> Result<CompilationResult, Error> {
+    let eir = translate_ast(ast)?;
+
+    compile_eir(TransformManager::default().apply(eir)?, module_name)
+}
+
+pub fn compile_string(source: &str, module_name: &str) -> Result<CompilationResult, Error> {
+    let ast = parser::parse(&source)?;
+    compile_ast(ast, module_name)
 }
