+++ src/compile/mod.rs
@@ -1,3 +1,4 @@
 pub mod codegen;
+pub mod compilation_result;
 pub mod error;
-pub mod llvm;
+pub mod target_machine;
